cache: check result == ngx.null when looking for Redis errors
resty.lua returns res, err for Redis operations. To know if there has been an
error accessing Redis, we need to check if res is null. When there is not an
error but we run an operation like 'get' with a key that does not exist,
resty.lua returns ngx.null instead of null. To differentiate from the error
case.

The tricky part is that redis-lua behaves differently. When there is an error
it simply executes error(). This commit modifies the tests accordingly
overwriting some methods to return ngx.null.