Rename xc.rockspec to apicast_xc.rockspec
This way we'll be able to differentiate between XC the library and the XC
module for Apicast (that uses the library internally).
This can be useful in the future for users who want to integrate the XC logic
in their projects but do not want to use Apicast or Openresty.