redis_pool: use XC_REDIS_HOST instead of REDIS_HOST
We need to rename the REDIS_HOST env var because Apicast is already using
REDIS_HOST and REDIS_PORT. Apicast only uses those env vars in the oauth flow.
I decided to use XC_REDIS_HOST instead of the same vars that Apicast uses to
give users more flexibility and allow them to use different storages for oauth
tokens and XC's cached auths/reports. Also, this gives the users the
possibility of scaling the XC storage when needed without changing the one that
stores the tokens and vice-versa.