Dockerfile-apicast: use apicast:v3.0.0-alpha2 as base image
We should not be using 'master' as the version of the base image. It's better to choose a version that we know for sure that works with XC. Unfortunately, there is not a 'final' version that contains all the changes that XC needs to function correctly. v3.0.0-alpha2 is the latest release and it works fine. Let's use it until v3.0.0 is released.