storage_keys: include app_key in report keys
If we do not include the app_key in the report key, the flusher will not be able to know which app keys it needs to renew the TTL for.
We were including other credentials already like app_id, user_key, access_token, etc. but we forgot about app_key.